feat: Convert api format to orc format
